+++ buildSrc/src/main/kotlin/versions.kt
@@ -1,9 +1,10 @@
 object vers {
-    val kotlin = "1.5.21"
+    val agp = "7.1.0-alpha05"
+    val kotlin = "1.5.10"
     val nexus_staging = "0.22.0"
 
     object kotlinx {
-        val coroutines = "1.5.1-native-mt"
+        val coroutines = "1.5.0-native-mt"
     }
 
     object androidx {
@@ -11,12 +12,13 @@ object vers {
     }
 
     object asoft {
-        val viewmodel = "0.0.92"
-        val live = "0.0.30"
-        val logging = "0.0.30"
-        val expect = "0.0.40"
+        val viewmodel = "0.0.93"
+        val live = "0.0.32"
+        val collections = "0.0.10"
+        val logging = "0.0.33"
+        val expect = "0.0.42"
+        val test = "1.1.40"
         val reakt = "0.1.10"
-        val builders = "1.3.40"
-        val test = "1.1.32"
+        val builders = "1.3.42"
     }
 }