+++ viewmodel-core/src/nativeMain/kotlin/viewmodel/PlatformViewModel.kt
@@ -1,5 +1,0 @@
-package viewmodel
-
-actual abstract class PlatformViewModel actual constructor() {
-    protected actual open fun onCleared() {}
-}